CA: implement calculating utilization for DRA resources
The logic is very basic and will likely need to be revised, but it's
something for initial testing. Utilization of a given Pool is calculated
as the number of allocated devices in the pool divided by the number of
all devices in the pool. For scale-down purposes, the max utilization
of all Node-local Pools is used.

The new logic is behind the DRA flag guard, so this should be a no-op
if the flag is disabled.