CA: move NodeInfo methods from SnapshotBase to ClusterSnapshot
All the NodeInfo methods have to take DRA into account, and the logic
for that will be the same for different SnapshotBase implementations.
Instead of duplicating the new logic in Basic and Delta, the methods
are moved to ClusterSnapshot and the logic will be implemented once in
PredicateSnapshot.

PredicateSnapshot will use the DRA Snapshot exposed by its SnapshotBase
to implement these methods. The DRA Snapshot has to be stored in the
SnapshotBase layer, as we need to be able to fork/commit/revert it.

Lower-level methods for adding/removing just the schedulerframework.NodeInfo
parts are added to SnapshotBase. PredicateSnapshot utilizes these methods
to implement AddNodeInfo and RemoveNodeInfo.

This should be a no-op, it's just a refactor.